requirements.txt
Django==1.11.29
django-crispy-forms==1.6.0
django-registration-redux==1.4
psycopg2
django-storages[azure]